+++ superset/datasets/commands/importers/v1/utils.py
@@ -1,0 +1,42 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+from typing import Any, Dict
+
+from sqlalchemy.orm import Session
+
+from superset.connectors.sqla.models import SqlaTable
+
+
+def import_dataset(
+    session: Session, config: Dict[str, Any], overwrite: bool = False
+) -> SqlaTable:
+    existing = session.query(SqlaTable).filter_by(uuid=config["uuid"]).first()
+    if existing:
+        if not overwrite:
+            return existing
+        config["id"] = existing.id
+
+    # should we delete columns and metrics not present in the current import?
+    sync = ["columns", "metrics"] if overwrite else []
+
+    # import recursively to include columns and metrics
+    dataset = SqlaTable.import_from_dict(session, config, recursive=True, sync=sync)
+    if dataset.id is None:
+        session.flush()
+
+    return dataset
