+++ superset/datasets/schemas.py
@@ -122,3 +122,48 @@ class DatasetRelatedDashboards(Schema):
 class DatasetRelatedObjectsResponse(Schema):
     charts = fields.Nested(DatasetRelatedCharts)
     dashboards = fields.Nested(DatasetRelatedDashboards)
+
+
+class ImportV1ColumnSchema(Schema):
+    column_name = fields.String(required=True)
+    verbose_name = fields.String()
+    is_dttm = fields.Boolean()
+    is_active = fields.Boolean(allow_none=True)
+    type = fields.String(required=True)
+    groupby = fields.Boolean()
+    filterable = fields.Boolean()
+    expression = fields.String()
+    description = fields.String(allow_none=True)
+    python_date_format = fields.String(allow_none=True)
+
+
+class ImportV1MetricSchema(Schema):
+    metric_name = fields.String(required=True)
+    verbose_name = fields.String()
+    metric_type = fields.String(allow_none=True)
+    expression = fields.String(required=True)
+    description = fields.String(allow_none=True)
+    d3format = fields.String(allow_none=True)
+    extra = fields.String(allow_none=True)
+    warning_text = fields.String(allow_none=True)
+
+
+class ImportV1DatasetSchema(Schema):
+    table_name = fields.String(required=True)
+    main_dttm_col = fields.String(allow_none=True)
+    description = fields.String()
+    default_endpoint = fields.String()
+    offset = fields.Integer()
+    cache_timeout = fields.Integer()
+    schema = fields.String()
+    sql = fields.String()
+    params = fields.String(allow_none=True)
+    template_params = fields.String(allow_none=True)
+    filter_select_enabled = fields.Boolean()
+    fetch_values_predicate = fields.String(allow_none=True)
+    extra = fields.String(allow_none=True)
+    uuid = fields.UUID(required=True)
+    columns = fields.List(fields.Nested(ImportV1ColumnSchema))
+    metrics = fields.List(fields.Nested(ImportV1MetricSchema))
+    version = fields.String(required=True)
+    database_uuid = fields.UUID(required=True)
